pyproject.toml
[project]
name = "fcbot"
version = "0.1.0"
description = "A FreeCAD Asset Generation Automation Tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.5",
    "pydantic>=2.10.6",
    "pydantic-core>=2.27.2",
    "pypdf>=5.3.0",
    "pyyaml>=6.0.2",
]
